workers/requirements.txt
openai>=1.55.0
azure-identity>=1.19.0
azure-servicebus>=7.12.2
azure-storage-blob>=12.22.0
python-dotenv>=1.0.0
